[core] add six and replace custom compat functions (#751)
* [core] vendor six and replace custom compat functions

* make_async_decorator

* Add docs about dependencies

* remove vendored 'six'
