+++ ddtrace/compat.py
@@ -1,5 +1,20 @@
 import platform
 import sys
+import textwrap
+
+import six
+
+__all__ = [
+    'httplib',
+    'iteritems',
+    'PY2',
+    'Queue',
+    'stringify',
+    'StringIO',
+    'urlencode',
+    'parse',
+    'reraise',
+]
 
 PYTHON_VERSION_INFO = sys.version_info
 PY2 = sys.version_info[0] == 2
@@ -8,32 +23,58 @@ PY2 = sys.version_info[0] == 2
 PYTHON_VERSION = platform.python_version()
 PYTHON_INTERPRETER = platform.python_implementation()
 
-stringify = str
-
-if PY2:
-    from urllib import urlencode
-    import httplib
-    stringify = unicode
-    from Queue import Queue
-    try:
-        from cStringIO import StringIO
-    except ImportError:
-        from StringIO import StringIO
-else:
-    from queue import Queue
-    from urllib.parse import urlencode
-    import http.client as httplib
-    from io import StringIO
-
 try:
-    import urlparse as parse
+    StringIO = six.moves.cStringIO
 except ImportError:
-    from urllib import parse
+    StringIO = six.StringIO
 
-try:
+httplib = six.moves.http_client
+urlencode = six.moves.urllib.parse.urlencode
+parse = six.moves.urllib.parse
+Queue = six.moves.queue.Queue
+iteritems = six.iteritems
+reraise = six.reraise
+
+stringify = six.text_type
+string_type = six.string_types[0]
+msgpack_type = six.binary_type
+# DEV: `six` doesn't have `float` in `integer_types`
+numeric_types = six.integer_types + (float, )
+
+
+if PYTHON_VERSION_INFO[0:2] >= (3, 4):
     from asyncio import iscoroutinefunction
-    from .compat_async import _make_async_decorator as make_async_decorator
-except ImportError:
+
+    # Execute from a string to get around syntax errors from `yield from`
+    # DEV: The idea to do this was stolen from `six`
+    #   https://github.com/benjaminp/six/blob/15e31431af97e5e64b80af0a3f598d382bcdd49a/six.py#L719-L737
+    six.exec_(textwrap.dedent("""
+    import functools
+    import asyncio
+
+
+    def make_async_decorator(tracer, coro, *params, **kw_params):
+        \"\"\"
+        Decorator factory that creates an asynchronous wrapper that yields
+        a coroutine result. This factory is required to handle Python 2
+        compatibilities.
+
+        :param object tracer: the tracer instance that is used
+        :param function f: the coroutine that must be executed
+        :param tuple params: arguments given to the Tracer.trace()
+        :param dict kw_params: keyword arguments given to the Tracer.trace()
+        \"\"\"
+        @functools.wraps(coro)
+        @asyncio.coroutine
+        def func_wrapper(*args, **kwargs):
+            with tracer.trace(*params, **kw_params):
+                result = yield from coro(*args, **kwargs)  # noqa: E999
+                return result
+
+        return func_wrapper
+    """))
+
+else:
     # asyncio is missing so we can't have coroutines; these
     # functions are used only to ensure code executions in case
     # of an unexpected behavior
@@ -44,20 +85,14 @@ except ImportError:
         return fn
 
 
-def iteritems(obj, **kwargs):
-    func = getattr(obj, "iteritems", None)
-    if not func:
-        func = obj.items
-    return func(**kwargs)
-
-
+# DEV: There is `six.u()` which does something similar, but doesn't have the guard around `hasattr(s, 'decode')`
 def to_unicode(s):
     """ Return a unicode string for the given bytes or string instance. """
     # No reason to decode if we already have the unicode compatible object we expect
-    # DEV: `stringify` will be a `str` for python 3 and `unicode` for python 2
+    # DEV: `six.text_type` will be a `str` for python 3 and `unicode` for python 2
     # DEV: Double decoding a `unicode` can cause a `UnicodeEncodeError`
     #   e.g. `'\xc3\xbf'.decode('utf-8').decode('utf-8')`
-    if isinstance(s, stringify):
+    if isinstance(s, six.text_type):
         return s
 
     # If the object has a `decode` method, then decode into `utf-8`
@@ -65,9 +100,9 @@ def to_unicode(s):
     if hasattr(s, 'decode'):
         return s.decode('utf-8')
 
-    # Always try to coerce the object into the `stringify` object we expect
+    # Always try to coerce the object into the `six.text_type` object we expect
     #   e.g. `to_unicode(1)`, `to_unicode(dict(key='value'))`
-    return stringify(s)
+    return six.text_type(s)
 
 
 def get_connection_response(conn):
@@ -86,45 +121,3 @@ def get_connection_response(conn):
         return conn.getresponse(buffering=True)
     else:
         return conn.getresponse()
-
-
-if PY2:
-    string_type = basestring
-    msgpack_type = basestring
-    numeric_types = (int, long, float)
-else:
-    string_type = str
-    msgpack_type = bytes
-    numeric_types = (int, float)
-
-if PY2:
-    # avoids Python 3 `SyntaxError`
-    # this block will be replaced with the `six` library
-    from .utils.reraise import _reraise as reraise
-else:
-    def reraise(tp, value, tb=None):
-        """Python 3 re-raise function. This function is internal and
-        will be replaced entirely with the `six` library.
-        """
-        try:
-            if value is None:
-                value = tp()
-            if value.__traceback__ is not tb:
-                raise value.with_traceback(tb)
-            raise value
-        finally:
-            value = None
-            tb = None
-
-
-__all__ = [
-    'httplib',
-    'iteritems',
-    'PY2',
-    'Queue',
-    'stringify',
-    'StringIO',
-    'urlencode',
-    'parse',
-    'reraise',
-]
