--- ddtrace/utils/reraise.py
@@ -1,5 +1,0 @@
-def _reraise(tp, value, tb=None):
-    """Python 2 re-raise function. This function is internal and
-    will be replaced entirely with the `six` library.
-    """
-    raise tp, value, tb
