+++ tox.ini
@@ -620,4 +620,7 @@ python_files = test*\.py
 
 [flake8]
 max-line-length=120
-exclude=.git,__pycache__,.tox,.ddtox,.eggs,*.egg
+exclude=
+  .ddtox,.tox,
+  .git,__pycache__,
+  .eggs,*.egg
